Merge in changes from wtsi-npg/master for version 0.8.0

+++ src/perl/t/database_pipeline.t
@@ -1,5 +1,5 @@
 
-# Tests WTSI::Genotyping::Database::Pipeline
+# Tests WTSI::NPG::Genotyping::Database::Pipeline
 
 use utf8;
 
@@ -12,18 +12,21 @@ use Test::Exception;
 
 use Data::Dumper;
 
-BEGIN { use_ok('WTSI::Genotyping::Schema'); }
-require_ok('WTSI::Genotyping::Schema');
+BEGIN { use_ok('WTSI::NPG::Genotyping::Schema'); }
+require_ok('WTSI::NPG::Genotyping::Schema');
 
-use WTSI::Genotyping::Database::Pipeline;
+use WTSI::NPG::Genotyping::Database::Pipeline;
 
 Log::Log4perl::init('etc/log4perl_tests.conf');
 
 my $ini_path = './etc';
 my $dbfile = 't/pipeline.db';
-unlink($dbfile);
 
-my $db = WTSI::Genotyping::Database::Pipeline->new
+if (-e $dbfile) {
+  unlink($dbfile);
+}
+
+my $db = WTSI::NPG::Genotyping::Database::Pipeline->new
   (name => 'pipeline',
    inifile => "$ini_path/pipeline.ini",
    dbfile => $dbfile);
@@ -32,13 +35,13 @@ ok(-e $dbfile);
 ok($db->disconnect);
 undef $db;
 
-$db = WTSI::Genotyping::Database::Pipeline->new
+$db = WTSI::NPG::Genotyping::Database::Pipeline->new
   (name => 'pipeline',
    inifile => "$ini_path/pipeline.ini",
    dbfile => $dbfile);
 ok($db, 'A database file reopened.');
 
-ok(WTSI::Genotyping::Database::Pipeline->new
+ok(WTSI::NPG::Genotyping::Database::Pipeline->new
    (name => 'pipeline',
     inifile => "$ini_path/pipeline.ini",
     dbfile => $dbfile,
@@ -56,7 +59,7 @@ dies_ok { $db->should_not_autoload_this_method->all }
 
 $db->populate;
 
-is($db->snpset->count, 26, 'The snpset dictionary');
+is($db->snpset->count, 27, 'The snpset dictionary');
 is($db->method->count, 5, 'The method dictionary');
 is($db->relation->count, 2, 'The relation dictionary');
 is($db->state->count, 12, 'The state dictionary',);
@@ -170,3 +173,6 @@ $failed_sample->update;
 
 $failed_sample = $db->sample->find({id_sample => $sample_id});
 ok(!$failed_sample->include, "Sample excluded after consent_withdrawn");
+
+# Clean up
+unlink($dbfile);
